While trying to connect to my 7.4 server on port 5440, I found some bugs. The NumToStr(double) function will return 544 for 5440.0
the function should look like this
...
if (result.Contains(wxT(".")))
    while (result.Right(1) == "0")
        result.RemoveLast();
if (result.Right(1) == ".")
    result.RemoveLast();

The call to NumToStr in pgConn.cpp (Line 91) should be a cast to long
instead of double.

Regards,

Andreas Pflug


git-svn-id: svn://svn.pgadmin.org/trunk@916 a7884b65-44f6-0310-8a51-81a127f17b15

+++ src/db/pgConn.cpp
@@ -88,7 +88,7 @@ pgConn::pgConn(const wxString& server, const wxString& database, const wxString&
     }
     if (port > 0) {
       connstr.Append(wxT(" port="));
-      connstr.Append(NumToStr((double)port));
+      connstr.Append(NumToStr((long)port));
     }
     connstr.Trim(FALSE);
